web: expose `setRenderAttributes` via a React hook
