+++ client/observability-client/src/react/index.ts
@@ -1,3 +1,3 @@
 export * from './TraceSpanProvider'
-export * from './useExistingSpan'
+export * from './useCurrentSpan'
 export * from './constants'
